Use uint32_t instead of long int for connection options.
Options should have a fixed width anyway, but this also fixes a possible MinGW
compiler bug where %lx tries to print a 64 bit value, even though a long int is
only 32 bits.

+++ src/connection.h
@@ -56,7 +56,7 @@ typedef struct connection_t {
 	int protocol_version;		/* used protocol */
 
 	int socket;					/* socket used for this connection */
-	long int options;			/* options for this connection */
+	uint32_t options;			/* options for this connection */
 	connection_status_t status;	/* status info */
 	int estimated_weight;		/* estimation for the weight of the edge for this connection */
 	struct timeval start;		/* time this connection was started, used for above estimation */
